# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.AppBriefConfig:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    required:
    - name
    type: object
...
  request.AppKnowledgebaseConfig:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/request.AppKnowledgebaseParams'
        description: 知识库参数
      knowledgebases:
        description: 知识库id、名字
        items:
          $ref: '#/definitions/request.AppKnowledgeBase'
        type: array
    type: object
...
  request.AppModelConfig:
    properties:
      config:
        description: 模型配置
      displayName:
        description: 模型展示名称(请求非必填)
        type: string
      examples:
        allOf:
        - $ref: '#/definitions/mp.AppModelParams'
        description: 仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数
      model:
        description: 模型名称
        type: string
      modelId:
        description: 模型ID
        type: string
      modelType:
        description: 模型类型(llm/embedding/rerank)
        type: string
      provider:
        description: 模型供应商
        type: string
    type: object
...
  request.AssistantBrief:
    properties:
      assistantId:
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    required:
    - assistantId
    - name
    type: object
  request.AssistantConfig:
    properties:
      assistantId:
        type: string
      instructions:
        description: 系统提示词
        type: string
      knowledgeBaseConfig:
        allOf:
        - $ref: '#/definitions/request.AppKnowledgebaseConfig'
        description: 知识库
      modelConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: 模型
      onlineSearchConfig:
        allOf:
        - $ref: '#/definitions/request.OnlineSearchConfig'
        description: 在线搜索
      prologue:
        description: 开场白
        type: string
      recommendQuestion:
        description: 推荐问题
        items:
          type: string
        type: array
      rerankConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: Rerank模型
      safetyConfig:
        allOf:
        - $ref: '#/definitions/request.AppSafetyConfig'
        description: 敏感词表配置
      visionConfig:
        allOf:
        - $ref: '#/definitions/request.VisionConfig'
        description: 视觉配置
    required:
    - assistantId
    type: object
  request.AssistantIdRequest:
    properties:
      assistantId:
        type: string
    required:
    - assistantId
    type: object
  request.AssistantMCPToolAddRequest:
    properties:
      actionName:
        type: string
      assistantId:
        type: string
      mcpId:
        type: string
      mcpType:
        enum:
        - mcp
        - mcpserver
        type: string
    required:
    - actionName
    - assistantId
    - mcpId
    - mcpType
    type: object
  request.AssistantMCPToolDelRequest:
    properties:
      actionName:
        type: string
      assistantId:
        type: string
      mcpId:
        type: string
      mcpType:
        enum:
        - mcp
        - mcpserver
        type: string
    required:
    - actionName
    - assistantId
    - mcpId
    - mcpType
    type: object
  request.AssistantMCPToolEnableRequest:
    properties:
      actionName:
        type: string
      assistantId:
        type: string
      enable:
        type: boolean
      mcpId:
        type: string
      mcpType:
        enum:
        - mcp
        - mcpserver
        type: string
    required:
    - actionName
    - assistantId
    - mcpId
    - mcpType
    type: object
...
  request.AssistantToolAddRequest:
    properties:
      actionName:
        type: string
      assistantId:
        type: string
      toolId:
        type: string
      toolType:
        enum:
        - builtin
        - custom
        type: string
    required:
    - actionName
    - assistantId
    - toolId
    - toolType
    type: object
  request.AssistantToolDelRequest:
    properties:
      actionName:
        type: string
      assistantId:
        type: string
      toolId:
        type: string
      toolType:
        enum:
        - builtin
        - custom
        type: string
    required:
    - actionName
    - assistantId
    - toolId
    - toolType
    type: object
  request.AssistantToolEnableRequest:
    properties:
      actionName:
        type: string
      assistantId:
        type: string
      enable:
        type: boolean
      toolId:
        type: string
      toolType:
        enum:
        - builtin
        - custom
        type: string
    required:
    - actionName
    - assistantId
    - toolId
    - toolType
    type: object
  request.AssistantWorkFlowAddRequest:
    properties:
      assistantId:
        type: string
      workFlowId:
        type: string
    required:
    - assistantId
    - workFlowId
    type: object
  request.AssistantWorkFlowDelRequest:
    properties:
      assistantId:
        type: string
      workFlowId:
        type: string
    required:
    - assistantId
    - workFlowId
    type: object
  request.AssistantWorkFlowToolEnableRequest:
    properties:
      assistantId:
        type: string
      enable:
        type: boolean
      workFlowId:
        type: string
    required:
    - assistantId
    - workFlowId
    type: object
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.ConversationCreateRequest:
    properties:
      assistantId:
        type: string
      prompt:
        type: string
    required:
    - assistantId
    - prompt
    type: object
  request.ConversationIdRequest:
    properties:
      conversationId:
        type: string
    required:
    - conversationId
    type: object
...
  request.ConversionStreamRequest:
    properties:
      assistantId:
        type: string
      conversationId:
        type: string
      fileInfo:
        items:
          $ref: '#/definitions/request.ConversionStreamFile'
        type: array
      prompt:
        type: string
      trial:
        type: boolean
    required:
    - assistantId
    - prompt
    type: object
...
  response.Assistant:
    properties:
      assistantId:
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      createdAt:
        description: 创建时间
        type: string
      desc:
        description: 描述
        type: string
      instructions:
        description: 系统提示词
        type: string
      knowledgeBaseConfig:
        allOf:
        - $ref: '#/definitions/request.AppKnowledgebaseConfig'
        description: 知识库
      mcpInfos:
        description: MCP信息
        items:
          $ref: '#/definitions/response.MCPInfos'
        type: array
      modelConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: 模型
      name:
        description: 名称
        type: string
      onlineSearchConfig:
        allOf:
        - $ref: '#/definitions/request.OnlineSearchConfig'
        description: 在线搜索配置
      prologue:
        description: 开场白
        type: string
      recommendQuestion:
        description: 推荐问题
        items:
          type: string
        type: array
      rerankConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: Rerank模型
      safetyConfig:
        allOf:
        - $ref: '#/definitions/request.AppSafetyConfig'
        description: 敏感词表配置
      scope:
        description: 作用域
        type: integer
      toolInfos:
        description: 自定义工具、内置工具
        items:
          $ref: '#/definitions/response.ToolInfos'
        type: array
      updatedAt:
        description: 更新时间
        type: string
      visionConfig:
        allOf:
        - $ref: '#/definitions/response.VisionConfig'
        description: 视觉配置
      workFlowInfos:
        description: 工作流信息
        items:
          $ref: '#/definitions/response.WorkFlowInfos'
        type: array
    required:
    - assistantId
    - name
    type: object
  response.AssistantCreateResp:
    properties:
      assistantId:
        type: string
    type: object
...
  response.ConversationCreateResp:
    properties:
      conversationId:
        type: string
    type: object
  response.ConversationDetailInfo:
    properties:
      assistantId:
        type: string
      conversationId:
        type: string
      createdAt:
        type: integer
      createdBy:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      prompt:
        type: string
      qa_type:
        type: integer
      requestFiles:
        items:
          $ref: '#/definitions/response.RequestFile'
        type: array
      response:
        type: string
      searchList: {}
      sysPrompt:
        type: string
      updatedAt:
        type: integer
    type: object
  response.ConversationInfo:
    properties:
      assistantId:
        type: string
      conversationId:
        type: string
      createdAt:
        type: string
      title:
        type: string
    type: object
...
  response.MCPInfos:
    properties:
      actionName:
        type: string
      enable:
        type: boolean
      mcpId:
        type: string
      mcpName:
        type: string
      mcpType:
        enum:
        - mcp
        - mcpserver
        type: string
      uniqueId:
        type: string
      valid:
        type: boolean
    required:
    - mcpType
    type: object
...
  response.PageResult:
    properties:
      list: {}
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- |
    --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n|
    X-Org-Id      | 组织ID    |\r\n| X-Client-Id   | 客户端标识|\r\n\r\n## HTTP Status\r\n|
    HTTP Status             | 说明                   |\r\n| -----------------------
    | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n|
    400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败
    \           |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n|
    一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n|
    guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      |
    【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n|
    permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      |
    权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n##
    `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n
    \   \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\":
    \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n
    \       {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n
...
    \"操作成功\"\r\n}\r\n```"
...
  version: v0.0.1
paths:
...
  /assistant:
    get:
      consumes:
      - application/json
      description: 查看智能体详情
      parameters:
      - description: 智能体id
        in: query
        name: assistantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.Assistant'
              type: object
      security:
      - JWT: []
      summary: 查看智能体详情
      tags:
      - agent
    post:
      consumes:
      - application/json
      description: 创建智能体，填写基本信息，创建完成为草稿状态
      parameters:
      - description: 智能体基本信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AppBriefConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AssistantCreateResp'
              type: object
      security:
      - JWT: []
      summary: 创建智能体
      tags:
      - agent
    put:
      consumes:
      - application/json
      description: 修改智能体基本信息，名称，头像，简介
      parameters:
      - description: 智能体基本信息参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantBrief'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 修改智能体基本信息
      tags:
      - agent
  /assistant/config:
    put:
      consumes:
      - application/json
      description: 修改智能体配置信息，模型配置，知识库配置等等
      parameters:
      - description: 智能体配置信息参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 修改智能体配置信息
      tags:
      - agent
  /assistant/conversation:
    delete:
      consumes:
      - application/json
      description: 删除智能体对话
      parameters:
      - description: 智能体对话的id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ConversationIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除智能体对话
      tags:
      - agent
    post:
      consumes:
      - application/json
      description: 创建智能体对话
      parameters:
      - description: 智能体对话创建参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ConversationCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ConversationCreateResp'
              type: object
      security:
      - JWT: []
      summary: 创建智能体对话
      tags:
      - agent
  /assistant/conversation/detail:
    get:
      consumes:
      - application/json
      description: 智能体对话详情历史列表
      parameters:
      - description: 智能体对话id
        in: query
        name: conversationId
        required: true
        type: string
      - description: 页面编号，从1开始
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 单页数量，从1开始
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ConversationDetailInfo'
                        type: array
                    type: object
              type: object
      security:
      - JWT: []
      summary: 智能体对话详情历史列表
      tags:
      - agent
  /assistant/conversation/list:
    get:
      consumes:
      - application/json
      description: 智能体对话列表
      parameters:
      - description: 智能体id
        in: query
        name: assistantId
        required: true
        type: string
      - description: 页面编号，从1开始
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 单页数量，从1开始
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ConversationInfo'
                        type: array
                    type: object
              type: object
      security:
      - JWT: []
      summary: 智能体对话列表
      tags:
      - agent
  /assistant/copy:
    post:
      consumes:
      - application/json
      description: 复制智能体，创建一个新的智能体，基本信息和配置都和原智能体一致
      parameters:
      - description: 智能体id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AssistantCreateResp'
              type: object
      security:
      - JWT: []
      summary: 复制智能体
      tags:
      - agent
  /assistant/stream:
    post:
      consumes:
      - application/json
      description: 智能体流式问答
      parameters:
      - description: 智能体流式问答参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ConversionStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 智能体流式问答
      tags:
      - agent
...
  /assistant/tool:
    delete:
      consumes:
      - application/json
      description: 为智能体解绑自定义、内建工具
      parameters:
      - description: 智能体id与自定义、内建工具id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantToolDelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除自定义、内建工具
      tags:
      - agent
    post:
      consumes:
      - application/json
      description: 为智能体绑定自定义、内建工具
      parameters:
      - description: 自定义、内建工具新增参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantToolAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 添加自定义、内建工具
      tags:
      - agent
  /assistant/tool/mcp:
    delete:
      consumes:
      - application/json
      description: 为智能体解绑mcp
      parameters:
      - description: mcp工具id、mcp类型、智能体id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantMCPToolDelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除mcp
      tags:
      - agent
    post:
      consumes:
      - application/json
      description: 为智能体绑定已发布的mcp工具
      parameters:
      - description: mcp工具id、mcp类型、智能体id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantMCPToolAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 添加mcp工具
      tags:
      - agent
  /assistant/tool/mcp/switch:
    put:
      consumes:
      - application/json
      description: 修改智能体绑定的MCP的启用状态
      parameters:
      - description: mcp工具id、mcp类型、智能体id、enable
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantMCPToolEnableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 启用/停用 MCP
      tags:
      - agent
  /assistant/tool/switch:
    put:
      consumes:
      - application/json
      description: 修改智能体绑定的自定义、内建工具的启用状态
      parameters:
      - description: 智能体id与自定义、内建工具id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantToolEnableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 启用/停用自定义、内建工具
      tags:
      - agent
  /assistant/tool/workflow:
    delete:
      consumes:
      - application/json
      description: 为智能体解绑工作流
      parameters:
      - description: 工作流id,智能体id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantWorkFlowDelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除工作流
      tags:
      - agent
    post:
      consumes:
      - application/json
      description: 为智能体绑定已发布的工作流
      parameters:
      - description: 工作流新增参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantWorkFlowAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 添加工作流
      tags:
      - agent
  /assistant/tool/workflow/switch:
    put:
      consumes:
      - application/json
      description: 修改智能体绑定的工作流的启用状态
      parameters:
      - description: 工作流id,智能体id,开关
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssistantWorkFlowToolEnableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 启用/停用工作流
      tags:
      - agent
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"